i#3544 core, part 10: Add various RISC-V stubs
Add a minimal implementation of symbols required to get the core of
DynamoRIO compiling.
This changeset contains code from multiple modules and was not separated
as it would create merge dependencies for code charing the same file.
Additionally those definitions are stubs for compilation purposes, so
there doesn't seem to be much to be gained from the separation.

NOTE:
  This code is not validated and mostly contains stubs as the main point
  was to achieve compilation. Some of the trivial logic has been
  implemented though. Proper implementation of stubs will follow in
  later commits.

Issue: #3544

Signed-off-by: Stanislaw Kardach <kda@semihalf.com>